+++ infrastructure/docker-compose.deploy.yml
@@ -400,7 +400,6 @@ services:
       - ES_PASSWORD=${ELASTICSEARCH_SUPERUSER_PASSWORD}
     volumes:
       - '/opt/opencrvs/infrastructure/monitoring/elastalert/elastalert.yaml:/opt/elastalert/config.yaml'
-      - '/opt/opencrvs/infrastructure/monitoring/elastalert/auth.yaml:/opt/elastalert/auth.yaml'
       - '/opt/opencrvs/infrastructure/monitoring/elastalert/rules:/opt/elastalert/rules'
     networks:
       - overlay_net
@@ -567,7 +566,7 @@ services:
     deploy:
       labels:
         - 'traefik.enable=true'
-        - 'traefik.http.routers.countryconfig.rule=Host(`countryconfig.{{hostname}}`)'
+        - 'traefik.http.routers.countryconfig.rule=Host(`countryconfig.{{hostname}}`) && !Path(`/email`)'
         - 'traefik.http.services.countryconfig.loadbalancer.server.port=3040'
         - 'traefik.http.routers.countryconfig.tls=true'
         - 'traefik.http.routers.countryconfig.tls.certresolver=certResolver'
@@ -579,6 +578,11 @@ services:
         - 'traefik.http.middlewares.countryconfig.headers.stsseconds=31536000'
         - 'traefik.http.middlewares.countryconfig.headers.stsincludesubdomains=true'
         - 'traefik.http.middlewares.countryconfig.headers.stspreload=true'
+        # This is an invalid IP range, effectively blocking all IPs from accessing /email path.
+        # It's only meant to be accessed from the internal docker network.
+        - 'traefik.http.middlewares.block-email.ipwhitelist.sourcerange=255.255.255.255'
+        - 'traefik.http.routers.block-email.rule=Host(`countryconfig.{{hostname}}`) && Path(`/email`)'
+        - 'traefik.http.routers.block-email.middlewares=block-email'
       replicas: 1
     environment:
       - MONGO_URL=mongodb://mongo1/user-mgnt?replicaSet=rs0
@@ -587,6 +591,16 @@ services:
       - COUNTRY_CONFIG_URL=https://countryconfig.{{hostname}}
       - LOGIN_URL=https://login.{{hostname}}
       - CLIENT_APP_URL=https://register.{{hostname}}
+      - NOTIFICATION_TRANSPORT=${NOTIFICATION_TRANSPORT}
+      - ALERT_EMAIL=${ALERT_EMAIL:-""}
+      - SMTP_HOST=${SMTP_HOST:-""}
+      - SMTP_PORT=${SMTP_PORT:-""}
+      - SMTP_USERNAME=${SMTP_USERNAME:-""}
+      - SMTP_PASSWORD=${SMTP_PASSWORD:-""}
+      - SMTP_SECURE=${SMTP_SECURE:-""}
+      - INFOBIP_GATEWAY_ENDPOINT=${INFOBIP_GATEWAY_ENDPOINT:-""}
+      - INFOBIP_API_KEY=${INFOBIP_API_KEY:-""}
+      - INFOBIP_SENDER_ID=${INFOBIP_SENDER_ID:-""}
       - DOMAIN={{hostname}}
     networks:
       - overlay_net
