+++ infrastructure/docker-compose.production-deploy.yml
@@ -96,12 +96,14 @@ services:
       - OPENHIM_URL=http://openhim-core:5001/fhir
       - CONFIRM_REGISTRATION_URL=http://openhim-core:5001/confirm/registration
       - CHECK_INVALID_TOKEN=true
-      - EMAIL_API_KEY=${EMAIL_API_KEY}
-      - INFOBIP_GATEWAY_ENDPOINT=${INFOBIP_GATEWAY_ENDPOINT:-""}
-      - INFOBIP_API_KEY=${INFOBIP_API_KEY:-""}
-      - INFOBIP_SENDER_ID=${INFOBIP_SENDER_ID:-""}
-      - SENDER_EMAIL_ADDRESS=${SENDER_EMAIL_ADDRESS}
       - SENTRY_DSN=${SENTRY_DSN}
+      - SENDER_EMAIL_ADDRESS=${SENDER_EMAIL_ADDRESS}
+      - ALERT_EMAIL=${ALERT_EMAIL}
+      - SMTP_HOST=${SMTP_HOST}
+      - SMTP_PORT=${SMTP_PORT}
+      - SMTP_USERNAME=${SMTP_USERNAME}
+      - SMTP_PASSWORD=${SMTP_PASSWORD}
+      - SMTP_SECURE=${SMTP_SECURE}
     deploy:
       replicas: 2
