+++ infrastructure/monitoring/elastalert/rules/alert.yaml
@@ -21,13 +21,6 @@ timestamp_type: 'iso'
 use_strftime_index: false
 # https://github.com/Yelp/elastalert/issues/1260
 raw_count_keys: false
-alert_subject: '{{HOST}} {}: {} 🚨'
-alert_subject_args:
-  - 'kibana.alert.context.metrics__alert__inventory__threshold.alertState'
-  - 'rule.name'
-alert_text: 'Reason: {}. Login to https://kibana.{{HOST}} to view the alert.'
-alert_text_args:
-  - 'kibana.alert.context.metrics__alert__inventory__threshold.reason'
 filter:
   - bool:
       should:
@@ -42,12 +35,10 @@ filter:
               value: 'Low on available disk space'
       minimum_should_match: 1
 
-alert:
-  - 'email'
-email: '{{ALERT_EMAIL}}'
-from_addr: 'alert@opencrvs.org'
-smtp_host: '{{SMTP_HOST}}'
-# prettier-ignore
-smtp_port: {{SMTP_PORT}}
-smtp_ssl: false
-smtp_auth_file: '/opt/elastalert/auth.yaml'
+alert: post2
+http_post2_url: 'http://countryconfig:3040/email'
+http_post2_payload:
+  subject: '{% raw %}{{DOMAIN}}{% endraw %} {{kibana.alert.context.metrics__alert__inventory__threshold.alertState}}: {{rule.name}} 🚨'
+  html: 'Reason: {{kibana.alert.context.metrics__alert__inventory__threshold.reason}}. Login to https://kibana.{% raw %}{{DOMAIN}}{% endraw %} to view the alert.'
+  from: '{% raw %}{{SENDER_EMAIL_ADDRESS}}{% endraw %}'
+  to: '{% raw %}{{ALERT_EMAIL}}{% endraw %}'
