+++ package.json
@@ -38,15 +38,14 @@
     "@octokit/core": "4.2.1",
     "@types/bcryptjs": "^2.4.2",
     "@types/google-libphonenumber": "^7.4.23",
-    "@types/hapi__inert": "5.2.1",
     "@types/handlebars": "^4.1.0",
+    "@types/hapi__inert": "5.2.1",
     "@types/jsonwebtoken": "^8.5.8",
     "@types/libsodium-wrappers": "^0.7.10",
     "@types/react-intl": "^3.0.0",
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
     "cypress-xpath": "^2.0.1",
-    "dotenv": "^6.1.0",
     "eslint": "^8.43.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
@@ -65,11 +64,11 @@
     "ts-jest": "^26.5.6"
   },
   "dependencies": {
+    "dotenv": "^6.1.0",
     "@faker-js/faker": "^6.0.0-alpha.5",
     "@hapi/boom": "^9.1.1",
     "@hapi/hapi": "^20.0.1",
     "@hapi/inert": "^6.0.3",
-    "@sendgrid/mail": "^7.7.0",
     "@types/chalk": "^2.2.0",
     "@types/code": "^4.0.3",
     "@types/csv2json": "^1.4.0",
@@ -86,6 +85,7 @@
     "@types/mime-types": "^2.1.0",
     "@types/node": "^10.12.5",
     "@types/node-fetch": "^2.6.2",
+    "@types/nodemailer": "^6.4.14",
     "@types/pino": "^5.14.0",
     "@types/string-similarity": "^3.0.0",
     "@types/uuid": "^3.4.4",
@@ -110,6 +110,7 @@
     "lodash": "^4.17.21",
     "mongoose": "6.11.3",
     "node-fetch": "^2.6.1",
+    "nodemailer": "^6.9.8",
     "opener": "^1.5.1",
     "p-queue": "^6.6.2",
     "pino": "^5.14.0",
